packages/middlewares/image-processors/blip2-image-processor/src/container/requirements.txt
boto3==1.34.22
transformers==4.36.2
Pillow==10.2.0
torchvision==0.16.2
